requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.4
python-dotenv==1.0.1
Werkzeug==3.0.1
scapy==2.5.0
pyfiglet==1.0.2
tqdm==4.66.2